+++ Backend/GroupManagement/src/main/java/com/vti/dto/booking/BookingFormForUpdating.java
@@ -9,7 +9,7 @@ public class BookingFormForUpdating {
 	private short tripId;
 	private short userId;
 	private short numOfPeople;
-	private BigDecimal totalPrice;
+	private int totalPrice;
 	private String details;
 
 	public short getTripId() {
@@ -36,11 +36,11 @@ public class BookingFormForUpdating {
 		this.numOfPeople = numOfPeople;
 	}
 
-	public BigDecimal getTotalPrice() {
+	public int getTotalPrice() {
 		return totalPrice;
 	}
 
-	public void setTotalPrice(BigDecimal totalPrice) {
+	public void setTotalPrice(int totalPrice) {
 		this.totalPrice = totalPrice;
 	}
